feat: added search functionality
In this commit I added made a search state which i gave to Search component as prop
and set it based on input value. Everytime user types, the search will be updated
and a debounce of 500ms delay will occur. The debounce function is a custom useEffect
hook that delays the query by 500ms in the Movies component. The debounce value will be
used as query dependency and search value in the URL.

MUST FIX: When input is empty, error shows Error: [movies,] data is undefined.
it should show something like 'Search movies by typing in the searchbox' or something