chore: used filterdialogcontext to handle dialog visibility
In this commit I used the isOpen state of FilterDialogContext to set the filter dialog's open and close
state. I removed the filter button as dialogtrigger in the dialog component and instead placed the
filter button in the discovermovies and only open it when isOpen is true. Also, instead of fetching the movies and all genres using useQueries, I got back to using separate useQuery for both of them since it is much cleaner.