feat: added searchfilterscontext
In this commit, I added a search filters context which contains an array of all filters based on the
selected filters in the filter dialog. I changed the genrefilter from isSelected to filters.include(genre)
to check if that filter is selected. I also moved all contexts to Providers component to avoid messy
context nesting. I also updated all usequery staletimes to Infinity (for faster loading times since I
observed that the movies do not change that much.