feat: genre filters are now added to url when clicking confirm
In this commit I got rid of SelectedFiltersContext and used a local state to store all filters in
filter dialog. I added a confirm button in filter dialogs so that when you select filters and click
confirm, all filters will be added to with_genres= in url and if no filters are selected, with_genres
will be removed in url.