fix: fixed apply same in different order issue
issue : when filters are [1,2,3] then remove filters to make [], then apply filters again but this time [2,3,1], it goes to else and removes with_genres params.
Expected Behavior:

- prev = [1,2,3], turn into [], then apply [3,1,2] => should just setIsOpen(false) since array content is the same

I fixed this issue by setting these conditions:

if filters.length!==0 and arrays are unequal, this means there are filters, but current filters are different from prev filters, so apply current filters

if filters.length!==0 and arrays are equal, this means there are filters, and prev and current filters are the same, so just close the dialog to avoid unnecessary fetch.

if filters.length===0 and arrays are unequal, this means no filters are applied, but searchparamsfilter with_genres has filters, this means user wants to clear filters.