feat: movies refetch when filters change
In this commit, I added filters (with_genres in URL) as a dependency in useFetchMovies
and returning movies with filters when there are filters applied. In the FilterDialog component,
I changed URLParamsFilterArray to make an array using searchParams instead of
ew URLSearchParams
since if I add the params as a dependency, it will cause an infinite loop since a new URLSearchParams
will be made everytime. I used searchParams since it is read-only. I also added isOpen to the
dependency array so that when user changes filters but chooses not to confirm, when the user opens
the dialog again, the filters state are reset back to the filters in the URL.