feat: clearing search input when deleting query in params
To clear the search input when query no longer exists in the URL, I made a state for the value of the
search input and set it on keystroke. I made a useEffect that runs when searchParams change, and sets the searchinput to the query in URL if it exists, otherwise sets it to empty string.